requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic-settings
python-dotenv
psycopg2-binary
zarinpal-py-sdk
